test(pg): the search-brain FTS language tests require the test database instead of skipping (#878)
The file read OPENBRAIN_TEST_DATABASE_URL itself and swapped in describe.skip
when it was unset, so a run without a database reported 0 pass / 14 skip and
exited 0 -- indistinguishable at the exit code from three suites that ran and
passed, while the Postgres behavior nobody exercised drifted underneath it.

All three live suites now call requireTestDatabaseUrl() at module scope. Absent
the variable the helper throws test_database_required and the run fails loudly.

The file also carried three independent live suites in 552 lines, so it is split
one suite per file, each with its own module-scope pool ended in its own
afterAll:

- search-brain-fts-language.pg.test.ts keeps the ranking suite.
- search-brain-fts-language-fields.pg.test.ts holds the migration-007
  source-field coverage.
- search-brain-fts-language-config.pg.test.ts holds the declared-language
  explicit-request chain.
- search-brain-fts-language-test-helpers.ts holds the shared module-scope
  helpers (assertUtf8Harness, embed, THOUGHTS_ONLY). It creates no pool and
  holds no test; assertUtf8Harness takes the pool as its first parameter.

Test bodies, names, and assertions moved verbatim; only the describe wrappers,
imports, pool creation, and helper plumbing changed. The three describe names
are unchanged, so all three scripts/assert-db-tests-ran.ts entries keep their
lines. JUnit confirms each emitted count still matches the recorded minimum
(4 / 1 / 3), and MIN_TOTAL_LIVE_TESTCASES is therefore unchanged at 265.

Each file's per-suite fixtures and query helpers move from inside the describe
callback to module scope alongside the pool they already used, which is what
keeps the three describe callbacks within the repo oxlint function-length rule
now that the pre-commit gate lints each staged file whole. The `pool as any` /
`deps as any` casts the original carried become `pool as never` (the existing
repo pattern) and a real `ToolDeps` annotation, clearing no-explicit-any in the
same pass. No assertion, test name, or tool under test changed.

Done-means: scripts/done-means/878-pg-tests-require-database.sh
